+++ kibot/kicad/worksheet.py
@@ -315,7 +315,7 @@ class WksText(WksDrawing):
             if e.font.italic:
                 s.SetItalic(True)
             if e.rotate:
-                s.SetTextAngle(e.rotate*10)
+                s.SetTextAngle(GS.angle(e.rotate))
             # Adjust the text size to the maximum allowed
             if e.max_len > 0:
                 w = s.GetBoundingBox().GetWidth()
@@ -381,7 +381,7 @@ class WksPolygon(WksDrawing):
                 s.SetWidth(e.line_width)
                 s.SetLayer(p.layer)
                 if e.rotate:
-                    s.Rotate(GS.p2v_k7(pos), e.rotate*10)
+                    s.Rotate(GS.p2v_k7(pos), GS.angle(e.rotate))
                 p.board.Add(s)
                 p.pcb_items.append(s)
             pos += posi