+++ kibot/out_base.py
@@ -885,12 +885,13 @@ class VariantOptions(BaseOptions):
         return fname
 
     @staticmethod
-    def save_tmp_dir_board(id, force_dir=None):
+    def save_tmp_dir_board(id, force_dir=None, forced_name=None):
         """ Save the PCB to a temporal dir.
             Disadvantage: all relative paths inside the file becomes useless
             Aadvantage: the name of the file remains the same """
         pcb_dir = mkdtemp(prefix='tmp-kibot-'+id+'-') if force_dir is None else force_dir
-        fname = os.path.join(pcb_dir, GS.pcb_basename+'.kicad_pcb')
+        basename = forced_name if forced_name else GS.pcb_basename
+        fname = os.path.join(pcb_dir, basename+'.kicad_pcb')
         logger.debug('Storing modified PCB to `{}`'.format(fname))
         GS.board.Save(fname)
         pro_name = GS.copy_project(fname)